+++ .gitignore
@@ -3,3 +3,4 @@ frontend/node_modules
 frontend/dist
 backend/__pycache__
 learnblocks/__pycache__
+*.swp
